+++ netlink-request/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "netlink-request"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+netlink-sys = "0.8"
+netlink-packet-core = "0.4"
+netlink-packet-generic = "0.2"
+netlink-packet-route = "0.10"
+netlink-packet-wireguard = "0.1"
